Rename `utils` to `pvsc_utils` to prevent shadowing
